# ADIOS: use OpenMP to speedup for loops