add `#pragma omp parallel for` to long for loop

~~**Do not merge before #907**~~