@psychocoderHPC 2015-06-03
Performance comparing (related to #858)

Setup:
- Simulation example: KHI
- one GPU
- 2 * 26*10^6 particles
- no ADIOS-IO (set transport layer to `NULL`)

**current dev**
- 11,6 seconds per species (only kernel `copySpecies`)

**new method without OpenMP**
`copy mallocMC` is only called one time at the beginning of the simulation
```
malloc  2sec  49msec = 2.04942 sec
copy mallocMC 674msec = 0.674744 sec
copy supercell e -  0msec = 4.49219e-05 sec
copy supercell i -  0msec = 5.29785e-05 sec
host copy kernel e - 169msec = 0.169465 sec
host copy kernel i - 167msec = 0.16786 sec
```

**new method *with* OpenMP**
`copy mallocMC` is only called one time at the beginning of the simulation
```
malloc  1sec 772msec = 1.77251 sec
copy mallocMC 657msec = 0.657336 sec 6246.68MiB/s
copy supercell e -  0msec = 3.80859e-05 sec
copy supercell i -  0msec = 0.000150146 sec
host copy kernel e - 172msec = 0.17208 sec
host copy kernel i - 172msec = 0.172543 sec
```