@ax3l 2015-06-04
haha, small change set and great effect :dancers: 

we should remember to add the individual environment variable [OMP_NUM_THREADS](https://gcc.gnu.org/onlinedocs/libgomp/Environment-Variables.html) to our `tpl` files in a follow-up.